% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-ROC.R
\name{create_survivalROC_data}
\alias{create_survivalROC_data}
\title{\code{create_survivalROC_data()} function}
\usage{
create_survivalROC_data(tvec, data, marker, time = time, status = status)
}
\description{
\code{create_survivalROC_data()} function
}
