% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cvaglmnet-parms.R
\name{get_cvaglmnet_params}
\alias{get_cvaglmnet_params}
\title{optimal cva (hyper)parameters}
\usage{
get_cvaglmnet_params(fit)
}
\description{
Extract the optimal (hyper)parameters from \code{cva.glmnet} object
}
