% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-cv-mytidy.R
\name{myglance.cv.glmnet}
\alias{myglance.cv.glmnet}
\title{\itemize{
\item @ templateVar class cv.glmnet
\item @ template title_desc_glance
}}
\usage{
\method{myglance}{cv.glmnet}(x, ...)
}
\description{
\itemize{
\item @inherit tidy.cv.glmnet params examples
}
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}

Other glmnet tidiers: 
\code{\link{mytidy.cv.glmnet}()}
}
\concept{glmnet tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
  \item{alpha}{Estimated correlation parameter for geepack::geeglm.}
  \item{lambda.1se}{The value of the penalization parameter lambda that results in the sparsest model while remaining within one standard error of the minimum loss.}
  \item{lambda.min}{The value of the penalization parameter lambda that achieved minimum loss as estimated by cross validation.}
  \item{nobs}{Number of observations used.}

}
