% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetUtils-cva-mytidy.R
\name{mytidy.cva.glmnet}
\alias{mytidy.cva.glmnet}
\title{importFrom purrr map_dfr}
\usage{
\method{mytidy}{cva.glmnet}(x, return_zeros = FALSE, unnest = TRUE, alpha_info = FALSE, ...)
}
\description{
importFrom purrr map_dfr
}
