% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-mytidy.R
\name{mytidy.glmnet}
\alias{mytidy.glmnet}
\title{-@templateVar class glmnet}
\usage{
\method{mytidy}{glmnet}(x, return_zeros = FALSE, component = c("coef", "dev"), ...)
}
\arguments{
\item{x}{A \code{glmnet} object returned from \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}

\item{`return_zeros`}{Logical indicating whether coefficients with value zero
zero should be included in the results. Defaults to \code{FALSE}.
\itemize{
\item @template param_unused_dots
}}
}
\description{
-@template title_desc_tidy
}
\details{
Note that while this representation of GLMs is much easier
to plot and combine than the default structure, it is also much
more memory-intensive. Do not use for large, sparse matrices.

No \code{augment} method is yet provided even though the model produces
predictions, because the input data is not tidy (it is a matrix that
may be very wide) and therefore combining predictions with it is not
logical. Furthermore, predictions make sense only with a specific
choice of lambda.
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{dev.ratio}{Fraction of null deviance explained at each value of lambda.}
  \item{estimate}{The estimated value of the regression term.}
  \item{lambda}{Value of penalty parameter lambda.}
  \item{step}{Which step of lambda choices was used.}
  \item{term}{The name of the regression term.}
  \item{step.label}{Label used by `glmnet`. It refers to column name in a matrix of coefficients. See example.}
  \item{df.step}{The number of nonzero coefficients for each value of lambda. For multnet, the number
     of variables with a nonzero coefficient for any class.}

}
