% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.penAFTFormula.R
\name{penAFT.cva}
\alias{penAFT.cva}
\title{Elastic net cross-validation of alpha and lambda for penAFT model}
\usage{
penAFT.cva(x, ...)
}
\description{
Do elastic net cross-validation of alpha and lambda simultaneously for penAFT models by calling penAFT::penAFT.cv() function
}
