% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred-censored.R
\name{pred_censored_coxnet}
\alias{pred_censored_coxnet}
\title{Predicted values for censored:::_coxnet model}
\usage{
pred_censored_coxnet(
  model_fit,
  new_data,
  time_var = time,
  status_var = status,
  time_points = NULL
)
}
\description{
Invokes \code{censored:::predict._coxnet()} to calculate different types of predicted values for #' _coxnet model
}
