% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins21.R
\docType{data}
\name{proteins21}
\alias{proteins21}
\title{Data with expression of 21 proteinss}
\format{
A data frame with x rows and ? variables:
\describe{
\item{group}{Label1}
\item{score}{Label2}
}
}
\source{
\url{https://www.github.com/agalecki-work/penAFTutils}
}
\usage{
proteins21
}
\description{
??.
}
\keyword{datasets}
