% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsag-package.R
\docType{package}
\name{utilsag-package}
\alias{utilsag}
\alias{utilsag-package}
\title{utilsag: Datasets and utility functions.}
\description{
Datasets and utilty functions
}
\details{
For a complete list of functions use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ library(help = "utilsag")
}\if{html}{\out{</div>}}

or

\if{html}{\out{<div class="sourceCode">}}\preformatted{help.start()
}\if{html}{\out{</div>}}

and locate \code{utilsag} package
\itemize{
\item ??@importFrom glmnetUtils minlossplot
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www-personal.umich.edu/~agalecki/}
}

}
\keyword{internal}
