% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-cv-mytidy.R
\name{myglance.cv.glmnet}
\alias{myglance.cv.glmnet}
\title{\itemize{
\item @ templateVar class cv.glmnet
\item @ template title_desc_glance
}}
\usage{
\method{myglance}{cv.glmnet}(x, ...)
}
\description{
\itemize{
\item @inherit tidy.cv.glmnet params examples
}
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
}
\concept{glmnet tidiers}
